v6.6.3 / TODO #279 (DONE): ZKP-NL's Go verifier capped at n = 32, a type limit the other three did not share
ZKP_NL_MAX_N is 64 in C, Python and Java and was 32 in Go -- a TYPE limit, not
a policy: the statement was a uint32 and the shares [3]uint32, so Go's decoders
rejected a wider key or proof before the verifier saw it, and an artifact the
other three produce came back as "n out of range".

Settled the two questions #279 said to answer before touching code:

  * The wire is UNAFFECTED.  Every ZKP-NL field -- A, B, y, a view's share and
    out-share -- is written nb = ceil(n/8) bytes wide rather than at a fixed
    four, so the encoding is width-tagged by the header's n.  No stored
    artifact changes and none becomes unreadable.  Go library callers do see
    uint32 -> uint64 on five exported functions.
  * The scope is PARITY at 64, not a wider family.  64 is C's declaration, not
    a derived security level, and #278 deliberately raised no widths.

The guard came first, as #279 asked: an n = 64 section in
CliTest/test_zkp_hybrid_family.sh -- the two 4x4 nl-zkboo / nl-zkbpp matrices
plus four-way pkey --pubout -- failed 9 cells against the pre-fix Go binary,
every one of them Go's, and passes 133/0 after.  Its key is generated by Python
because only Python and Java can produce one; that Go asymmetry is filed as
#281.

Go's own test [22] had "proved" ZKBoo at n = 64 throughout, hollowly: A and B
were truncated into a uint32 and uint32((1 << 64) - 1) is all-ones, so both of
its widths carried 32 significant bits.  Only a cross-language check could see
this.

zkpNlMask(n) replaces six copies of (1 << n) - 1 with an explicit n >= 64
branch -- at exactly the width this change exists to reach, a 64-bit shift of a
uint64 is 0 and the mask would come out right by accident.  The aPAKE witness
(HpakeRecord.B/.Y) moved with the type, since it feeds ZkpNlProve directly.

PARAM_DIVERGENCE's zkp-nl-max-n row is deleted as the orphan rule forces,
leaving spec/'s parameter axis with zero defect rows.

+++ HerraduraCli/herradura_cli.go
@@ -2477,7 +2477,7 @@ func decodeZkpRnlProof(body []byte) (w, c, z []int, n int, err error) {
 }
 
 // encodeZkpNlPriv: Wire: 4B n | nb A | nb B | nb y
-func encodeZkpNlPriv(A, B, y uint32, n int) string {
+func encodeZkpNlPriv(A, B, y uint64, n int) string {
 	nb := (n + 7) / 8
 	buf := make([]byte, 4+3*nb)
 	binary.BigEndian.PutUint32(buf[0:], uint32(n))
@@ -2493,7 +2493,7 @@ func encodeZkpNlPriv(A, B, y uint32, n int) string {
 	return PemWrap(lblZkpNlPriv, buf)
 }
 
-func decodeZkpNlPriv(body []byte) (A, B, y uint32, n int, err error) {
+func decodeZkpNlPriv(body []byte) (A, B, y uint64, n int, err error) {
 	if len(body) < 4 {
 		return 0, 0, 0, 0, fmt.Errorf("ZKP-NL privkey too short")
 	}
@@ -2506,19 +2506,19 @@ func decodeZkpNlPriv(body []byte) (A, B, y uint32, n int, err error) {
 		return 0, 0, 0, 0, fmt.Errorf("ZKP-NL privkey truncated")
 	}
 	for i := 0; i < nb; i++ {
-		A = (A << 8) | uint32(body[4+i])
+		A = (A << 8) | uint64(body[4+i])
 	}
 	for i := 0; i < nb; i++ {
-		B = (B << 8) | uint32(body[4+nb+i])
+		B = (B << 8) | uint64(body[4+nb+i])
 	}
 	for i := 0; i < nb; i++ {
-		y = (y << 8) | uint32(body[4+2*nb+i])
+		y = (y << 8) | uint64(body[4+2*nb+i])
 	}
 	return
 }
 
 // encodeZkpNlPub: Wire: 4B n | nb B | nb y
-func encodeZkpNlPub(B, y uint32, n int) string {
+func encodeZkpNlPub(B, y uint64, n int) string {
 	nb := (n + 7) / 8
 	buf := make([]byte, 4+2*nb)
 	binary.BigEndian.PutUint32(buf[0:], uint32(n))
@@ -2531,7 +2531,7 @@ func encodeZkpNlPub(B, y uint32, n int) string {
 	return PemWrap(lblZkpNlPub, buf)
 }
 
-func decodeZkpNlPub(body []byte) (B, y uint32, n int, err error) {
+func decodeZkpNlPub(body []byte) (B, y uint64, n int, err error) {
 	if len(body) < 4 {
 		return 0, 0, 0, fmt.Errorf("ZKP-NL pubkey too short")
 	}
@@ -2544,10 +2544,10 @@ func decodeZkpNlPub(body []byte) (B, y uint32, n int, err error) {
 		return 0, 0, 0, fmt.Errorf("ZKP-NL pubkey truncated")
 	}
 	for i := 0; i < nb; i++ {
-		B = (B << 8) | uint32(body[4+i])
+		B = (B << 8) | uint64(body[4+i])
 	}
 	for i := 0; i < nb; i++ {
-		y = (y << 8) | uint32(body[4+nb+i])
+		y = (y << 8) | uint64(body[4+nb+i])
 	}
 	return
 }
